refactor(Promotion Popup): Promoter extracted to be usable out of popover
